Fix issue with the Dispose method
Depending on backends to implement IDisposable is problematic, since
a backend may be a native widget class implementing IDisposable on its
own, so calling Dispose in this case would actually dispose the widget,
not the backend. IWidgetBackend now has a Dispose method but it
doesn't inherit from IDisposable anymore, so a backend can have
a Dispose implementation for the backend, independent from the
one from the widget.