Implemented new approach for calculating the preferred size of a widget
The calculation of the desired size of a WPF widget is now routed through
the frontend, by calling the OnGetPreferred* methods if they are overriden.
This is done by subclassing every widget and calling a common MeasureOverride
method provided by WidgetBackend. In this way, WPF's DesiredSize is now a
reliable measure, and can be used to calculate the size of widget containers,
without having to implement container-specific size calculation algorithms, since
we can just use what WPF provides.