Fix calculation of the default size of windows
When the window is shown for the first time and no specific size has been set,
use the natural size of the root widget as size for the window. Also ensure
that the window can't be shrinked beyond the min size of the root widget.

Also fixed the calculation of window non-client area. The correct property
for the caption size is WindowCaptionHeight.