ProgressBar: improve API
As adviced by Lluis, a new property Indeterminate is added,
to make it more understandable on what mode of the widget is
being used (instead of a Nullable Fraction property).

Also, as the default value for a bool is false, and it is
better to start the GLib.Timeout loop when indicated by the
user (by changing its Indeterminate value to true) in the
Gtk backend, now Indeterminate is false by default, and the
initial fraction is 0.0.