Add ForeignPlug widget
This widget serves as a barrier between toolkit, only bridge toolkit (i.e. GtkMac) implements it with whatever specific mechanism they have to host a toolkit widget hierarchy into another toolkit widget hierarchy.

In practise that means the parent and the child of a ForeignPlug are from two different toolkits.