[Xwt.Wpf] Allow the user to handle the event before handling internally
We need to attach an empty event handler so the virtual method is always invoked
and from there we launch the default action if the user has not handled the event
and performed a custom action.

Updated the samples so we can trivially check this behaviour is consistent across
toolkits.