[Mac] Rewrite to use CoreGraphics for Context and CoreText for TextLayout
The CGContext API is very similar to Cairo, enables us to set the global alpha,
and is probably more performant than calling the Objective-C drawing routines in AppKit.
Also use CoreText for TextLayout because it works with a CGContext and enables us to wrap
text correctly based on the given width.

Basically, the Mac drawing samples look much more like the Gtk version now.