[WPF] Fix window size calculation bug
In WPF, the size of the client area depends on the size of the window borders,
and the border size changes if the window is resizable or fixed. Now we
use the correct border size depending on the size mode. Also, we now ensure
the size of the client area doesn't change when the size mode changes.