[Mac] Fix ContextBackendHandler implementations of TransformPoint(s) and TransformDistance(s)
The issue was that the matrix returned by CGContextGetCTM includes the full view transform,
but we only care about the view-relative transform applied to the context. Fix that by
saving the initial CTM of the context and concatenating its inverse whenever we need to get the
context transform. This should be safe as the initial affine transform returned by CoreGraphics
for an NSView should never be singular. I think.