Associate the active Toolkit with the queued Task in the XwtTaskScheduler
This seems like the right thing to do because the active Toolkit is held
in a static field. Even though xwt should not be called from multiple threads,
there is still the possibility that the active toolkit may change between the
time the Task is queued and when it is actually run.