[Xwt.Mac] Make this csproj compilable via XBuild
First I made this a regular .NET library project and not a monomac
project. I then removed the import for the non-existent monomac
targets file. Finally I added the 'ReferencePath' attribute so that
if we cannot find 'monomac' in the Gac we will then look for it
in a parallel checkout of monomac.

With this changes Xwt.Mac still compiles as normal in MonoDevelop
but when compiled from the commandline as part of MonoDevelops build
we can now succesfully compile this and have it as part of the sln