Fix size calculation issue
Widgets that require a size calculation were stored in order of
depth, since the calculation has to be done bottom-up. However,
this was not being properly done since the depth registered when
the resize request was made may have changed when the calculation
is done. Now, the widgets are ordered by depth just before doing
the size calculation.