[WPF] Fix Canvas rendering issues
Don't substract the margin from child allocation (margin is not included in
the children bounds).
Don't set the background of the WPF Canvas. The background is rendered in
the default implementation of OnRender, which we need to call after we do
our own rendering, since it also renders the child widgets. So if we set
the background, it is drawn over the custom rendering.