[WPF] Fix window sizing issue
The GetSystemMetrics API is known to return incorrect values for window
frame size metrics in some cases, due to a workaround in the Windows API
to support legacy applications running on Aero.
Instead of using that API to calculate the client area of the window,
we now rely on the size returned by the content. This requires some tricks
to force the window to adapt to content size changes.