Do not capture escape key when search text entry is empty.
Capturing the escape key prevents a dialog from closing.
When there is text in the search text entry then the escape key is
captured and the text is cleared when this key is pressed.
When there is no text in the search text entry then the escape key
is not handled/captured by the search text entry and can be
handled elsewhere, such as by a dialog.