[WPF] Resize bitmap before drawing if necessary.
When an image is a single bitmap that doesn't have the same intrinsic size as the drawing size, dc.DrawImage makes a very poor job of down/up scaling it.

Thus we handle this manually by using a TransformedBitmap to handle the conversion in a better way when it's needed.