Revert "[WPF] Removed support for default text foreground, which was interfering with region foregrounds."
This introduce a regression when drawing normal text and using ctx.SetColor by forcing the color to always be black.

Although the commit did fix an issue, the only way to do it properly and take into account the current context-set color is to defer the creation of the TextFormatted instance until drawing time so that the context-bound color can be passed to TextFormatted constructor as it can't be set afterwards.

This reverts commit db439534ae067b84646b613cde20d837bfc007ea.