[Xwt.Gtk] Handle when multiple drag/drop datatypes are allowed
If SetDragDropTarget is called and multiple types are passed in,
DragDataRequests will be set equal to the number of items passed
in.

Once we process all of those we need to emit the event. This is
done even if the drag operation did not contain any of the
supported transfer data types.

The exact case where this was failing was if you registered as supporting
Foo and Bar and then tried to do a drag of type Bar. On my system it
would check the 'Bar' TransferDataType' first and then not emit the
events because DragDataRequests was equal to 1. Then it woudl check
'Foo', which was not part of the drag and so there was no data, so
it would bail out before emitting.