[Gtk] add RenderWidget coordinates check
Some Gtk Widgets use an additional Gdk.Window for drawing
and/or their bounds do not match their GdkWindow window. A workaround
is to check if the window and widget allocation sizes are equal.
In that case the widget allocation offset could be relative to a higher level
window and should be ignored.