[Mac] fix Button sizing
The default NSBezelStyle.Rounded style (which is the default for text
buttons on Mac)  does not support cusom heights.
When a custom size or image is used, the style should be set to
NSBezelStyle.RegularSquare (the default for image or custom buttons
on Mac) for the sizing to work as expected.