[Xwt] Ensure TextLayout is properly initialized when using Toolkit ctor.
Since TextLayout inherits from XwtObject, the default ctor of the later is always called which sets the ToolkitEngine property to the current engine.

This breaks the logic of InitForToolkit which expect a null value for that property to detect the initialization case. So we reset that property value first before calling InitForToolkit.

The use case for this ctor is when recording drawing commands for VectorImage.