[Gtk] Don't force backing EventBox to have a background.
If using Canvas and having a non-square drawing. Enabling mouse events for instance forced the background of the eventbox to be shown which would result in white artifacts.

Instead we only show the box as necessary the same way it's done in the eventbox wrapper path.

Additionally, even if an event box has already been created, if the method is called again with its parameter set to true we propagate that information so that drawing can happen normally.