[Xwt.Mac] Don't crash if the TextEntry is destroyed while inside a KeyUp
There are two changes here:
1) We run our internal logic before letting the rest of the toolkit handle
the KeyUp event. This is so we always emit the event when the cursor position
changes, regardless of whether or not the user destroys the widget inside the
KeyUp handler they attached.

2) We null check 'CurrentEditor' and only access it if it's non-null. This
prevents a crash when destroying the editor inside the KeyUp handlers and
matches what the rest of the code does.