[Xwt.XamMac] Fix TextEntry for this backend
The 'GotFocus' and 'LostFocus' events should be implemented by
checking when the editor gets/loses focus. In order to do this
we need to create a custom field editor. This also allows us to
properly track the KeyDown/KeyUp events. The NSTextField itself
will not reliably get these events as they'll be consumed by the
editor.