[Xwt.XamMac] Gross hack to fix focusing for Xwt.TextEntry (again!)
It turns out we cannot rely on the base implementation of SetFocus
because the 'Widget' property is a different object in our subclass.

This means if we want focus to work, we have to override all the methods
and make sure we invoke the same code but using the correct object.

This makes focus work again.