[Xwt.Wpf] Properly handle TargetTypes
The 'FillDataStore' method uses the 'DataTypes' value to
figure out what values need to be copied over to the data store
which is given to the destination object. If we don't store these
values, no data is copied from the source store to the dest store
and drag/drop fails!

Fixes the ios designer on wpf mode when dragging from the toolbox
to the surface.