[Mac] Fix MacClipboardBackend
The 'CanReadItemWithDataConformingToTypes' method does not do what we
want. If we add a custom object to the clipboard and then invoke
'CanReadItemWithDataConformingToTypes' will return false.

It looks like the intent of the code is to check if a specific UTI
is available in the pasteboard, as such changing this to just look
in the 'Types' array should be sufficient as it will contain every
registered UTI.

Fixes copy/paste in the designer and bug
https://bugzilla.xamarin.com/show_bug.cgi?id=31827