[Mac] Add error handling to track down an inexplicable NullReferenceException
Somehow `source` is ending up as null. There are a few possible causes:
1) Xamarin.Mac bug where objects are not being retained correctly. There
was a bug in the past where NSCell.Copy was broken and the cloned object
would re-use the same GCHandle as the original object, which lead to bugs.

2) 'User' bug where the Cells are not being retained correctly and the wrapper
class is being collected early. Maybe we should be retaining it in managed
code somewhere.

3) Xwt bug and we're passing in a null `source` somewhere

We can test for the 3rd case by adding some handy exception throwing. If
they trigger then the issue is probably case 1 or 2 and we'll need more
diagnostics.