[Gtk] Disable BackSpace keybinding in Gtk.TreeView so that we can get key pressed event generated from it.
On Mac, the backspace key (actually labeled "delete") can be used for delete situations. The issue with treeview is that this key is by default mapped to the "select-cursor-parent" event which prevent the normal key pressed/released machinery from happening.

Since that same event is also bound to Ctrl+Backspace, we simply disable the binding for plain Backspace so that this key can be handled normally.