Add an option to choose how wrapped native widgets are sized
The original behavior was to return Size.Zero from
EmbeddedNativeWidget.OnGetPreferredSize. This works when the
size is explicitly set on the resultant Xwt.Widget, or when
the parent widget provides sizing. However, there are cases
(e.g. in a popover) where you need to size the parent based
on the preferred size of the child.

This patch keeps the original behavior as the default, but
allows you to opt in to having the default Widget backend
try to determine the preferred size of the wrapped widget.

This is an opt-in rehash of b93128864ddbd31e942cf3835b8a0604d946c735