[WPF] Use VisualTreeHelper to check for a widget parent.
Using the Parent property on a FrameworkElement only retrieves the logicial WPF parent of a widget which may not exists. That doesn't prevent the widget from being hosted in a visual hierarchy like on Visual Studio where its hosted in a special COM component.

Thus we switch the code to use VisualTreeHelper which correctly take care of those cases and can return a non-null parent in that case.

This commit also fixes a bunch of sizing issue on WPF when widgets are embedded in native toolkits because checks that relied on HasNativeParent would wrongly fail (like in the `NotifySizeChangeToParent` method).