Fix font toolkit validation for system fonts
When a system font is being validated and initialized
for a different toolkit, only its family name should be
checked whether its a system font or not and the actual
settings should be copied from the source backend.
Otherwise a new font with the original system font
family was created, which is not necessarily the system
font of the target toolkit and which may not be even
available for it (i.e. San Francisco fonts on OSX, which
is not supported by Pango).

(fixes bug #38321)