[Wpf] Improve TooltipText implementation to work better with Gtk classic workflow.
This patch does a couple of things:

- Fix NRE when querying TooltipText when no tooltip was set
- Use a real ToolTip instance in place of just setting a string.
- Make sure to hide tooltip when the value is set to null while the tooltip is showing (as is often the case with custom widget based on Canvas that set TooltipText while in a MouseEnter/MouseMove event)

This patch still doesn't solve some issue that works under toolkit. Notably, it still doesn't handle the fact that when the mouse is already on the widget and a tooltip is set in a mouse event for instance, the tooltip showing timer is not started until the mouse goes out and back in again. This seems to only appear once until the tooltip instance has been initialized.