[Xwt.Gtk] Add initial support for WPF embedding using MonoDevelop code.
This commit reuse the code developed in MonoDevelop[0] and adapt it for
xwt (for instance removing `IdeApp` references).

It works by embedding the WPF content contained in its own window handle
into our main Gtk.Window instance. It reserve space for it via the dummy
Gtk+ widget implementation so that it can participate into the layout and
forward focus events.

[0] https://github.com/mono/monodevelop/tree/master/main/src/core/MonoDevelop.Ide/MonoDevelop.Components/Windows