[Gtk] Optimize paragraph spacing in RichTextView
We use two line breaks for paragraphs (GtkTextView
does not distinguish between new lines and paragraphs,
which looks bad because of the full line space inbetween.

This commit adds a tag to the buffer which makes the
separating line 1px high with a 1/2 line spacing around it.
The spacing grows with the LineSpacing property.