[Gtk] Remove double subscription to SelectionChanged event.
The ListBox backend was adding its own handler to the underlying widget selection changed event using the base TableView sink. The issue is that the parent TableView does the same thing which results in the client code handler for selection changed to be called twice in a row for each single modification.