[Mac] Fix TreeView / ListView column autosizing
Cocoa does not support column autosizing
based on the table data. If AutosizingMask
is enabled, Cocoa will resize the columns
based on the header titles and on table
creation only.

If a column is not resizable by the user
(CanResize == false), we need to size it
manually to make the content of all rows
visible.