[Mac] Fix stack overflow in ViewBackend.ParentIsSensitive
In some rare cases a view and its superview both implement IViewObject
and additionally share the same Backend (for eaxmple a custom widget
inside an alignement container), which results in an infite loop when
Backend.ParentIsSensitive is called.

The workaround is to not rely on Superview being a IViewObject
and its Backend.Widget to be the real parent, but to use
Widget.Superview only to walk up the widget hierarchy.