[Mac] Optimize image drawing context.
This commit works around code that reflects a lot. NSObject.Copy() does
reflection to find the intptr constructor of an object, then uses
ConstructorInfo.Invoke to construct the NSImage.

Since we don't operate with the actual NSImage in the case we don't
have a custom image, do not create a copy, as we only need to pass in
the CGImage we get. The AsCGImage does not do any reflection, so this
should turn into faster code.