Perf opt (#694)
* [Xwt] Performance improvement on context handling.

Don't use a stack to handle the save/restore, but actually link the items together. This reduces allocations caused by the stacks on every operation to actually one more item allocation.
Also, normalize handling of save/restore. Do not throw on restore not having an item (this is made to be the same as Restore in Context.cs

* [Perf] Don't allocate lambda closures on drawing images.

* [Mac] CurrentStatus should not be null for the mac backend.

* Fixup! Missing part from previous commit