[Xwt] Don't double initialize a window.
IPopupWindowBackend should not care about IWindowFrameBackend initialization. Leave it to the popupwindow to handle that, if we do this from PopupWindowBackend, we end up initializing once from this handler, and another initialization being done in the invoker of IPopupWindowBackend.Initialize

This fixes a leak such as:
```
var popup = new Xwt.PopupWindow();
popup.Dispose();
```

Before, the objects would not even be marked for resource release.