[WPF] Fix measuring for CustomPanel
It is possible for IsArrangeValid to return true when
IsMeasureValid returns false. This is probably a bug
in WPF, but it is unfortunately what is happening.

This is a workaround for that issue. We should measure
when the measure is invalid :)

This fixes the case where the embedded exception dialog
or the loading spinner do not appear when opening a
storyboard file in Visual Studio.