[Mac] Ensure Copy(NSZone) is overriden
This ensures that if AppKit clones the native cell we will be able to
create a new C# wrapper for it.

One known cause of this is when calculating the BaselineOffsetFromBottom.
This issue also affected the designers.

This is being added as a preventative as we may not actually be
triggering an API which clones the native object, but it's better to
be safe than sorry!

See https://github.com/mono/monodevelop/pull/4465