[Wpf] Don't skip the Measure step all the time
If we add something to the live tree it will have
an invalid measure and arrange. The first measure pass
will run as-expected.

If you remove and re-add the widget to the live tree
then it's measure and arrange will be invalidated, but
because we do not call 'ArrangeChildren (force = true)'
then we will never invoke the Measure for the children
again.

For the case of ListBox this means an implciit measure
will occur, which passes Inf,Inf as the size, which
means we lose scrollbars as it expands to fill all
that space.

This fixes the ExceptionPopover in the designers.

(cherry picked from commit cd022e582ada54440bee219cb38200946191fb64)