[a11y][WPF] Fixed combo box item automation names
This fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/596863.
Previously, we showed Xwt.WPFBackend.ValueConverter (the class name) as the
automation Name for combo box items, but now properly show the item text as
the Name.

At first, I tried replacing the default ComboBox and ComboBoxItem automation peers,
but there was a fair amount of complexity around that, making the pair of peers
properly work together. In the end, overriding ToString turned out to be a much
simpler and safer fix.