[a11y][WPF] Fix accessibility focus handling for spin buttons
Previously, if a spin button had the focus, the narrator wouldn't see
it. We also reported to automation clients that spin buttons weren't
focusable. This fixes both of those issues. Now the spin button
having the focus is treated basically the same as the text box
inside the spinner control having the focus, as it's the text box
that gets focus when the user tabs to it.