[a11y][WPF] Added accessibilty support for popover windows
Previously, we weren't setting the automation Name on the window for
popovers, which this PR addresses, fixing:
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/594024
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/595199

Popovers are special for a couple of reasons:
- On the Xwt side, they derive from XwtComponent instead of XwtWidget.
- On the WPF side, accessibility info for the window should be set on the
PopupRoot, which doesn't exist until the popup is actually shown. Thus
we need to delay setting the accessibility settings until the popup is
shown.

These changes pass along accessbility settings for the non-widget
popovers, hold onto them, and delay setting them until the Show.