[WPF] Convert XAML styles to code to avoid UI stalling
Between the reflection, XAML parsing and inflating, a profiler shows that retrieving the style for the buttons can take close to a second of processing on the main thread causing a visible jank.

This commit converts that XAML code into its equivalent code so that we bypass those sources of slowness.