[VS] Support a11y info for popup menus, delay set
On WPF, XWT popup menus (ContentMenu objects) are created on demand
when first displayed. So we need to delay setting the a11y properties until
the menu is created.

Popup menus are similar to Popovers in this way, as they also delay setting
the a11y info until display time.