Make ThemedImage style selection less unpredictable
GetImage is optimized for performace and therefore just loops over
the images and selects the first best matching one based on given styles.
The result is not predictable if the image list is not sorted and several alternatives
match different global and explicit styles.
Since we want to prioritize explicit styles first, we need to make sure that we
don't select an image with a given global style but without an explicit match first.
The simplest approach without slowing down rendering by a sophisticated lookup is to
sort images matching global styles last, so that GetImage would always select
an alternative with an explicit rather than a global style.