Ignore NuGet pack warnings
Fix build on appveyor. NuGet now reports errors instead of
warnings for the licenseUrl and lib directory.

Error NU5125: The 'licenseUrl' element will be deprecated. Consider
using the 'license' element instead.
Error NU5128: Some target frameworks declared in the dependencies
group of the nuspec and the lib/ref folder do not have exact matches
in the other location. Consult the list of actions below:
- Add a dependency group for .NETFramework4.7.2 to the nuspec

Whilst the licenseUrl could be fixed the NU5128 cannot be fixed
and looks to be a bug in NuGet.

https://github.com/NuGet/Home/issues/8713

For now these errors are ignored when using nuget pack.