Properly handle escaping of markup text
Previously, Xwt assumed that the platform (Gtk) handled
encoding/decoding of markup text that contains special
characters. GLib.Markup.EscapeText would do the escaping
and then Gtk itself would unescape when displayed.

Since we switched from the Gtk to the Cocoa backend, that
has two problems - GLib.Markup.EscapeText crashes if
Gtk isn't there and even if it does work, there's nothing on
the Cocoa side to unescape the text.

Instead we now handle all escaping/unescaping in a backend
independent way. WebUtility.HtmlEncode does the escaping
and cross platform Xwt markup parsing code does the match
WebUtility.HtmlDecode to unescape.

For this to fully work the Xwt change and VSMac change need
to both be present, but the VSMac update is the most important
as it removes the crash.