Fix fatal Obj-C callback issue due to delegate collection
VS for Mac is encountering a fatal error of: A callback was made on a garbage collected delegate of type 'Xwt.XamMac!Xwt.Mac.ViewBackend+DelegateIntPtrIntPtrBool::Invoke'.

The underlying root cause for this is that the delegates that are created and passed off to native code are not being kept alive.  The GC has no way of knowing that delegates passed to native code need to be kept alive so a root must be established in some way.

Since these are all static methods, we can create a static variable to hold these delegates for the lifetime of the application.