Fix drag/drop support (#1091)
Fix so that drag/drop support works properly with the Cocoa backend.
This PR includes these fixes:

DraggingUpdate passed the wrong coordinates. Now do the proper conversion from view coordinates to window coordinates, as expected by the Xwt client.
Get the INSDraggingInfo object in the proper way, with Runtime.GetINativeObject<INSDraggingInfo> (dragInfo, owns: false), same as we do in other code now.
Use the newer NSView.BeginDraggingSession API instead of the deprecated NSView.DragImage. This is needed so that we can provide a DraggingSource object in order to be notified when the drag/drop is over (no matter if successful or not) and call the Finished callback, as expected by Xwt client code.
Populate the pasteboard item used for the drag properly, supporting text drag/drop and internal formats (normally a serialized object where the Id is the assembly qualified type name).
To my knowledge the Android Designer is the only client that still uses this drag/drop code. All of these fixes were to make Android Designer drag/drop work properly.

Squashed commits:

* Update InitPasteboard, including all data types
* Treat as internal data
* Update INSDraggingInfo type, thanks to Sandy
* More fixes
* Fix DraggingUpdated coordinates to be widget, not window
* Serialize drag data for Cocoa backend
This more closely matches what was
done on Gtk before. It’s needed for at
least the Android Designer, supporting
drag/drop on the AndroidDesignerDragData
serializable type.
* Use BeginDraggingSession for drag/drop
This allows sending the OnDragFinished notification
when the drag/drop is complete, in turn calling
DragOperation.Finished. Calling Finished is expected
for Xwt - it did before for Gtk and this makes
the Cocoa backend behave the same.
* Update pasteboard handling
* Only use valid names for pasteboard types
* Switch to using INSDraggingSource
* Spaces -> tabs