[Mac] Fix ImageView accessibility
The ImageView.Accessible was using the INSAccessibility implementation
from the NSImageView when it should have been using the NSImageView's
Cell.

Change the AccessibleBackend to use the NSImageView's Cell. The
ImageViewBackend puts the NSImageView inside an CustomAlignedContainer
so the AccessibleBackend.Initialize method checks for an NSImageView
after the CustomAlignedContainer.Child has been found.